pyproject.toml
[tool.poetry]
name = "convert"
version = "0.1.0"
description = "tools to convert files, ala csv -> json, yml -> json for localization of OTP-UI, etc..."
authors = ["Frank Purcell <fxpurcell@gmail.com>"]
license = "Mozella 2.x"
readme = "README.md"
packages = [{include = "ott/convert"}]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
mergedeep = "^1.3.4"

[tool.poetry.scripts]
gtfs_feeds = "ott.convert.csv2json.gtfs.feeds_to_json"

json2locs = "ott.convert.yaml2json.generate:json_to_localizations"
mockloc = "ott.convert.yaml2json.generate:mock_localization"
yml2json = "ott.convert.yaml2json.convert:yml_to_json"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
